Reorganize project and update on dialogs UI
