+++ frontend/src/components/common/NavBar.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import { useAuth } from "@/stores/authStore";
 import { AuthService } from "@/services/authService";
 import { useNavigate } from "react-router-dom";
@@ -27,6 +27,7 @@ const NavBar = () => {
     <header className="fixed top-0 left-0 right-0 z-50 border-b bg-card shadow-sm">
       <div className="container mx-auto flex h-16 items-center justify-between px-4">
         <div className="flex items-center gap-4">
+          <img src="/src/assets/mental-health-dark.png" alt="Alzheimer's Prediction System Logo" className="h-10 w-10" />
           <h1 className="text-xl font-bold">Alzheimer's Prediction System</h1>
         </div>
